Fix texture sync issue by sending plain resource paths and excluding .godot folder

+++ addons/team_create/scene_sync.gd
@@ -405,15 +405,19 @@ func _check_single_node_changes(node: Node):
 			if typeof(val) == TYPE_OBJECT:
 				# For resources like Mesh or Material, sync the resource path if possible
 				if val is Resource:
-					# Serialize local sub-resources or resources without a file path
-					var bytes = var_to_bytes_with_objects(val)
-					var temp_path = "user://tc_sync_export_" + str(val.get_instance_id()) + ".tres"
-					ResourceSaver.save(val, temp_path)
-					var text = ""
-					if FileAccess.file_exists(temp_path):
-						text = FileAccess.get_file_as_string(temp_path)
-						DirAccess.remove_absolute(temp_path)
-					current_props[p.name] = {"sub_resource_bytes": bytes, "sub_resource_text": text, "resource_path": val.resource_path}
+					if val.resource_path != "" and val.resource_path.begins_with("res://"):
+						# ONLY send the string path over the network
+						current_props[p.name] = val.resource_path
+					else:
+						# Serialize local sub-resources or resources without a file path
+						var bytes = var_to_bytes_with_objects(val)
+						var temp_path = "user://tc_sync_export_" + str(val.get_instance_id()) + ".tres"
+						ResourceSaver.save(val, temp_path)
+						var text = ""
+						if FileAccess.file_exists(temp_path):
+							text = FileAccess.get_file_as_string(temp_path)
+							DirAccess.remove_absolute(temp_path)
+						current_props[p.name] = {"sub_resource_bytes": bytes, "sub_resource_text": text, "resource_path": val.resource_path}
 			else:
 				current_props[p.name] = val
